+++ CHANGELOG.md
@@ -27,6 +27,7 @@ Unless otherwise specified, any version comparison below is the comparison of se
 - (API) Implemented writable book and written book, several classes were added. Check the commit history for details.
 - (API) Implemented (glowing) item frame, several classes were added. Check the commit history for details.
 - (API) Added event `ItemFrameUseEvent` which is called when an item frame is used.
+- (API) Added `scale` property for `DebugArrow`.
 - Added overloads for methods `AllayXXXType.Builder.build()` which accept a custom `ClassLoader`. This parameter should be used by plugins
   if plugins want to create custom item type, because plugins classes are loaded by `JarPluginClassLoader`.
 - Added support for in-plugin resource pack, which allows plugins to include their resource pack under `assets/resource_pack`. This